+++ src/Grids/extruded.jl
@@ -126,8 +126,8 @@ end
 
 topology(grid::ExtrudedFiniteDifferenceGrid) = topology(grid.horizontal_grid)
 
-is_continuous(grid::ExtrudedFiniteDifferenceGrid) =
-    is_continuous(grid.horizontal_grid)
+discretization(grid::ExtrudedFiniteDifferenceGrid) =
+    discretization(grid.horizontal_grid)
 
 ClimaComms.context(grid::ExtrudedFiniteDifferenceGrid) =
     ClimaComms.context(grid.horizontal_grid)